Crate gammalooprs

Modules

cff

feyngen

graph

initialisation

integrands

integrate

This module steers the integration process. It contains two main ways of integrating the integrand. The havana_integrate function is mostly used for local runs. The master node in combination with batch_integrate is for distributed runs.

model

momentum

numerator

observables

processes

settings

subtraction

tests

utils

uv

Macros

debug_cache

Helper macro for cache debugging

debug_cache_search

Helper macro for missed cache hit detection in debug mode

debug_tags

Shorthand for tagged debug events.

define_index

define_indexed_vec

disable

can be used instead of commenting out code

dot

monitor_cache

Helper macro for cache monitoring with custom conditions

warn_cache_efficiency

Helper macro for cache efficiency warnings

Structs

GammaLoopContextContainer

Enums

DependentMomentaConstructor

Constants

GAMMALOOP_NAMESPACE

MAX_CORES

MAX_LOOP

Statics

INTERRUPTED

ITERATION_ABORT_REQUESTED

Traits

GammaLoopContext

HasModel

LogMessage

Functions

clear_interrupt_request

is_interrupt_requested

request_interrupt

request_iteration_abort

set_interrupt_handler