gammalooprs::cff

Module esurface

Structs

Esurface

Core esurface struct

EsurfaceID

Index type for esurface, location of an esurface in the list of all esurfaces of a graph

ExistingEsurfaceId

Index in the list of all existing esurfaces, essentially a pointer to a pointer to an esurface

GroupEsurfaceId

RaisedEsurfaceData

RaisedEsurfaceGroup

RaisedEsurfaceId

Enums

EsurfaceExistenceStatus

Pointwise existence classification exposed without the internal diagnostic margin.

Type Aliases

EsurfaceCache

EsurfaceCollection

ExistingEsurfaces

Container for esurfaces that exist at a given point in the phase space

ExistingThresholds

ExternalShift