gammalooprs::graph

Struct Graph

pub struct Graph {
    pub overall_factor: Atom,
    pub name: String,
    pub group_id: Option<GroupId>,
    pub is_group_master: bool,
    pub tree_edges: SuBitGraph,
    pub underlying: HedgeGraph<Edge, Vertex, HedgeData>,
    pub loop_momentum_basis: LoopMomentumBasis,
    pub param_builder: ParamBuilder,
    pub global_prefactor: GlobalPrefactor,
    pub surface_cache: SurfaceCache,
    pub initial_state_cut: OrientedCut,
    pub polarizations: Vec<(PolDef, Atom)>,
}

Fields
overall_factor: Atomname: Stringgroup_id: Option<GroupId>is_group_master: booltree_edges: SuBitGraphunderlying: HedgeGraph<Edge, Vertex, HedgeData>loop_momentum_basis: LoopMomentumBasisparam_builder: ParamBuilderglobal_prefactor: GlobalPrefactorsurface_cache: SurfaceCacheinitial_state_cut: OrientedCut

The cross section initial state cut Only relevant for cross sections, but stored here for the parsing

polarizations: Vec<(PolDef, Atom)>
Implementations

impl Graph

pub fn cff<S: SubGraphLike + SubSetLike>( &mut self, contract_subgraph: &S, cutset: &CutSet, orientation_pattern: &OrientationPattern, ) -> Result<CutCFF>

impl Graph

pub fn to_dot_graph_with_settings( &self, settings: &DotExportSettings, ) -> DotGraph

pub fn to_split_dotgraph(&self) -> DotGraph

impl Graph

pub fn dot_serialize(&self, settings: &DotExportSettings) -> String

pub fn dot_serialize_fmt( &self, writer: &mut impl Write, settings: &DotExportSettings, ) -> Result<(), Error>

pub fn from_dot(graph: DotGraph, model: &Model) -> Result<Self>

pub fn from_file<P>(p: P, model: &Model) -> Result<Vec<Self>>
where P: AsRef<Path>,

pub fn from_path<P>(p: P, model: &Model) -> Result<Vec<Self>>
where P: AsRef<Path>,

pub fn from_string<Str: AsRef<str>>(s: Str, model: &Model) -> Result<Vec<Self>>

impl Graph

pub fn debug_dot(&self) -> String

pub fn debug_dot_with_settings(&self, settings: &DotExportSettings) -> String

pub fn pretty_dot(&self) -> String

pub fn get_edges_in_initial_state_cut(&self) -> Vec<EdgeIndex>

impl Graph

pub fn generate_polarizations(&self) -> Atom

pub fn generate_polarization_params(&self) -> Vec<Atom>