gammalooprs::integrands

Module evaluation

Structs

BatchSampleEvaluationResult

EvaluationMetaData

Useful metadata generated during the evaluation, this may be expanded in the future to include more information

EvaluationResult

The result of an evaluation of the integrand

EvaluationResultOutput

GenericEvaluationResult

GenericEvaluationResultOutput

GraphEvaluationResult

LoopMomentaEscalationMetrics

PreciseBatchSampleEvaluationResult

PreciseSampleEvaluationResult

PreciseSingleSampleEvaluationResult

RawBatchEvaluationResult

RawPreciseBatchEvaluationResult

RotatedEvaluation

SampleEvaluationResult

SingleSampleEvaluationResult

StabilityResult

StatisticsCounter

This struct merges the evaluation metadata of many evaluations into a single struct

Enums

PreciseEvaluationResult

PreciseEvaluationResultOutput

StabilityFailureReason

Per-precision evaluation details produced during stability checks.

StabilityStatus