gammalooprs::integrands

Module process

Re-exports

pub use sampling::CompiledSamplingChannel;

pub use sampling::CompiledSamplingMap;

pub use sampling::GaussianReferenceFunction;

pub use sampling::ImplicitSurfaceContextPreparer;

pub use sampling::ImplicitSurfaceRadialContextEvaluator;

pub use sampling::ImplicitSurfaceRadialEvaluator;

pub use sampling::ImplicitSurfaceRadialMap;

pub use sampling::PreparedSurfaceStatus;

pub use sampling::ReferenceMoments;

pub use sampling::ReferenceSampleEvaluation;

pub use sampling::ReferenceSamplingReport;

pub use sampling::ResolvedNamedSamplingChannel;

pub use sampling::ResolvedSamplingBlock;

pub use sampling::ResolvedSamplingChannelSelection;

pub use sampling::SamplingCatalogueEntry;

pub use sampling::SamplingChannelBridge;

pub use sampling::SamplingChannelBridgeAcceptanceReport;

pub use sampling::SamplingChannelBridgeError;

pub use sampling::SamplingChannelBridgeEvaluation;

pub use sampling::SamplingChannelCatalogue;

pub use sampling::SamplingChannelCompileContext;

pub use sampling::SamplingChannelCompileError;

pub use sampling::SamplingChannelId;

pub use sampling::SamplingChannelInspection;

pub use sampling::SamplingChannelPreset;

pub use sampling::SamplingChannelRuntimeContexts;

pub use sampling::SamplingChannelScore;

pub use sampling::SamplingChannelSelector;

pub use sampling::SamplingCoverageReport;

pub use sampling::SamplingCutSide;

pub use sampling::SamplingDualValue;

pub use sampling::SamplingExpressionEvaluator;

pub use sampling::SamplingGeometryKey;

pub use sampling::SamplingJacobian;

pub use sampling::SamplingMapAcceptanceReport;

pub use sampling::SamplingMapAffine;

pub use sampling::SamplingMapComponent;

pub use sampling::SamplingMapComposition;

pub use sampling::SamplingMapContextTransform;

pub use sampling::SamplingMapContract;

pub use sampling::SamplingMapDefinition;

pub use sampling::SamplingMapEmbedding;

pub use sampling::SamplingMapEvaluation;

pub use sampling::SamplingMapKernel;

pub use sampling::SamplingMapPoint;

pub use sampling::SamplingMomentumSampleContext;

pub use sampling::SamplingPartition;

pub use sampling::SamplingPartitionMode;

pub use sampling::SamplingScoreFunction;

pub use sampling::SamplingSelectionError;

pub use sampling::SamplingSupport;

pub use sampling::SharedEnergyJointGeometry;

pub use sampling::SharedEnergyJointMap;

pub use sampling::SurfaceRadialMap;

pub use sampling::SurfaceRadialPoint;

pub use sampling::build_sampling_channel_catalogue;

pub use sampling::build_sampling_channel_catalogue_with_surfaces;

pub use sampling::build_sampling_channel_catalogue_with_surfaces_and_coverage;

pub use sampling::explicitly_selected_graphs;

pub use sampling::graph_channel_definitions;

pub use sampling::resolve_sampling_channel_selection;

pub use sampling::resolve_sampling_channel_selection_replacing_default;

pub use evaluators::ActiveF64Backend;

pub use evaluators::GenericEvaluator;

pub use evaluators::GenericEvaluatorFloat;

pub use param_builder::ParamBuilder;

pub use param_builder::ParamValuePairs;

pub use param_builder::ThresholdParams;

pub use param_builder::UpdateAndGetParams;

Modules

amplitude

cache_debugging

Essential cache debugging utilities for GammaLoop

cross_section

evaluators

gammaloop_sample

ir

param_builder

sampling

Graph-aware sampling maps, their canonical channel catalogue, and validation.

threshold_multiplier

Structs

CacheStats

Cache usage statistics

CacheValidationResult

Result of cache validation checks

GraphTermEvaluationContext

LmbMultiChannelingSetup

Helper struct for the LMB multi-channeling setup

MomentumSpaceEvaluationInput

Direct input for evaluating an integrand at explicit spatial loop momenta.

Enums

EvaluationTarget

IntegrandType

OrientationProfileMode

ProcessIntegrand

Traits

GraphTerm

ProcessIntegrandImpl