gammalooprs

Module integrate

This module steers the integration process. It contains two main ways of integrating the integrand. The havana_integrate function is mostly used for local runs. The master node in combination with batch_integrate is for distributed runs.

Structs

BatchIntegrateInput

Input for the batch_integrate function, created by the master node

BatchResult

The result of evaluating a batch of points

ComplexAccumulator

HavanaIntegrateRequest

IntegralResult

IntegrationSlot

IntegrationState

struct to keep track of state, used in the havana_integrate function the idea is to save this to disk after each iteration, so that the integration can be resumed

IntegrationStatusViewOptions

IntegrationWorkspaceManifest

IterationBatchingSettings

MasterNode

Master node which accumulates data from jobs, and creates the input for jobs

RatatuiDashboardState

SerializableBatchIntegrateInput

SlotMeta

StatusUpdate

TabledRenderOptions

WorkspaceSnapshotControl

Enums

BatchIntegrateOutput

ContributionSortMode

EventOutput

Different ways of processing events, EventList is a list of events, Histogram does accumulation of events on the worker nodes, so the master node only has to merge the histograms.

EventOutputSettings

Choose whether to output an eventlist, a histogram, or nothing

IntegralOutputSettings

Choose whether to output the integrand results, or a statistics accumulator

IntegrationStatusKind

IntegrationStatusPhaseDisplay

SampleInput

Different ways of passing samples to the batch_integrate function One is simply a list of samples, the other is a grid and the desired number of samples The worker then generates the samples itself.

SamplingCorrelationMode

Functions

build_integration_result

emit_integration_status_via_tracing

havana_integrate

Integrate function used for local runs

latest_observable_resume_state_path

print_integral_result

render_saved_integration_summary

render_status_update_tabled

slot_workspace_path

workspace_manifest_path

workspace_result_snapshot_path

workspace_state_dir

workspace_state_path