gammalooprs::uv

Module profile

UV Profile Analysis Module

This module provides functionality for analyzing ultraviolet behavior of loop integrands by evaluating them at different momentum scalings and computing degrees of divergence.

Structs

Analysis

AnalyticAnalysis

AnalyticResult

InspectAnalysis

InspectResult

LMBResult

OrientationAnalyticAnalysis

ProfileSettings

SubSetResult

UVProfile

UVProfileAnalysis

UVProfileAnalyticEntry

UVProfileFailure

UVProfileFixedRay

UVProfileGraphAnalysis

UVProfileLmbAnalysis

UVProfileOrientationInspectEntry

UVProfilePassFail

UVProfileSubsetAnalysis

UVProfileSummary

UVSamplingResult

Traits

UVProfileable

Functions

logspace